Resolved issue #2

+++ src/main/java/dev/jacrispys/JavaBot/Audio/GuildAudioManager.java
@@ -347,7 +347,7 @@ public class GuildAudioManager {
 
         String pageNumber = "Page " + queuePage + "/" + (int) Math.ceil((float) scheduler.getTrackQueue().size() / 10);
         String trackInQueue = "Songs in Queue: " + trackList.size();
-        int queueLength = 0;
+        long queueLength = 0;
         String queueLengthStr;
         for (AudioTrack audioTrack : trackList) {
             queueLength += audioTrack.getDuration();
